Lift isWavSilent into separate utility.
